Overhauled key stroke handling, 1 bullet/space, turn + move @ same time

+++ src/hellofx/Bullet.java
@@ -11,11 +11,11 @@ public class Bullet extends Polyline {
     private double x, y, rotate;
     private double life;
     //private Rectangle size;
-    private int speed = 0;
+    private double speed = 0;
     long startTime = System.currentTimeMillis();
     public static final int BULLET_WIDTH = 2;
     
-public Bullet(double x1, double y1, double r1, int s) {
+public Bullet(double x1, double y1, double r1, double s) {
     // x1 and y1 to be nose position of ship that fired (either alien or space ship)
     super(Bullet.bulletPoints);
     this.setTranslateX(x1);
@@ -23,7 +23,7 @@ public Bullet(double x1, double y1, double r1, int s) {
     this.setRotate(r1);
     // setting to 6 so bullet travels at reasonable speed regardless
     // bullet speed dependent on ship speed
-    this.speed = 6 + s;
+    this.speed = 6.0 + s;
 
 }
 
@@ -56,7 +56,7 @@ public void wrap() {
 
 public void move() {
 	double playerAngle = this.getRotate();
-    int bulletSpeed = this.speed;
+    double bulletSpeed = this.speed;
     double currentX = this.getTranslateX();
     double currentY = this.getTranslateY();
     double currentVelocityX = bulletSpeed * Math.sin(Math.toRadians(playerAngle));
